Move Params to formatter/ dir
