+++ src/formatter/ExpressionFormatter.ts
@@ -1,7 +1,7 @@
 import type { FormatOptions } from 'src/types';
 import { equalizeWhitespace } from 'src/utils';
 
-import Params from 'src/core/Params';
+import Params from 'src/formatter/Params';
 import { isTabularStyle } from 'src/core/config';
 import { isReserved, type Token, TokenType } from 'src/core/token';
 import {
